requirements.txt
jsonpickle
tornado
click
git+https://github.com/johandahlberg/localq.git@with_shell_true # Get from pip in future - localq
mock # only for tests
pyyaml
illuminate
pandas

